✨[Feat] Home 마이그레이션 잔여 결선 — 참여자 선택·권한·섹션 에러·프로필 저장 (#1117)
* ✨[Feat] 홈 일정 등록 참여자 선택 결선 + ScheduleCapabilities 소비처 결선

일정 등록이 participantMemberIds 를 빈 배열로 하드코딩해 작성자 본인조차 참여자로
들어가지 않던 문제를 해결하고, DTO~UseCase~DI 등록까지 만들어두고 소비처가 0이던
ScheduleCapabilities 파이프라인을 화면에 결선한다.

참여자 선택
- SelectedChallengerView 와 init 을 public 으로 승격 (나머지 Challenger 컴포넌트는
  some View 뒤에 가려져 직접 참조되지 않으므로 internal 유지)
- Home/Project.swift presentationExtraDependencies 에 ActivityDomain·ActivityPresentation 추가
- 생성 시 작성자 memberId 를 항상 포함, 수정 시에는 변경된 경우에만 전송해 PATCH 3-state 유지
- 편집 진입 시 기존 참여자 프리필, EditFormSnapshot 에 participantMemberIds 추가

ScheduleCapabilities
- maxParticipantCount 를 Int → String 으로 교정 (절대 규칙 #2, 소비처 0인 지금이 최소 비용)
- .task 에서 loadCapabilities() 호출, 권한 없으면 출석 필수 토글 비노출
- 참여자 상한 초과 시 인라인 경고와 등록 차단, N / max 표시
- 조회 실패는 "권한 모름"으로 두어 기존 동작 유지. 수정 모드의 기존 출석 일정은 토글을
  강제로 내리지 않는다 (PATCH 가 해제로 나가 서버 출석 데이터가 지워짐)
- StubSession 에 ScheduleCapabilitiesRepositoryProtocol 등록 (누락 시 스텁 세션이 실네트워크 사용)

그 밖에
- AI 자동완성 실패를 전역 알럿 대신 시트 내 인라인 상태로 복구, 미지원 기기 안내 추가
- 피커 토글 시 키보드 내림, 제목 필드 Equatable 래퍼 복구
- ScheduleRegistrationViewModelTests 신규 12건

* 🐛[Fix] 이미 시작된 일정의 편집 진입 사전 차단

ScheduleDetailViewModel 의 isScheduleStarted 가 선언만 되고 어디서도 읽히지 않아,
사용자가 시작된 일정도 편집 화면을 끝까지 채운 뒤 서버(SCHEDULE-0028)에 거부당했다.

- isEditActionEnabled(canEditSchedule && !isScheduleStarted) 추가 후 툴바에 결선
- 항목별 disabled 와 접근성 label/hint 가 필요해 공용 ToolbarTrailingMenu 대신
  이 화면 전용 Menu 로 구성
- "이미 시작된 일정은 수정할 수 없습니다" 안내 배너 복원. 수정 권한이 없는 챌린저에게는
  설명할 진입점 자체가 없어 canEditSchedule && isScheduleStarted 로 게이팅
- 삭제 진행 중 메뉴가 조용히 비던 문제 수정 — 라벨 자리를 ProgressView 로 교체
- 등록 화면의 SCHEDULE-0028 처리는 최후 방어선으로 유지

* 🐛[Fix] 홈 섹션 에러·재시도 복원 및 홈 진입 프로필 저장 결선

최근 공지 섹션
- 조회 실패 시 EmptyView 를 반환해 헤더까지 사라지고 에러도 재시도도 없던 문제 수정.
  헤더를 상태 밖으로 빼 .failed 에서도 유지하고 RetryContentUnavailableView 로 대체
- 하드코딩돼 있던 isRetrying: false 를 실제 상태로 결선
- 재시도 시 기수 정보가 없으면 프로필부터 다시 조회
- 밋밋한 ProgressView 를 LoadingView(.home(...)) 로 교체

홈 진입 프로필 저장
- SyncProfileStorageUseCase(Protocol) 를 AuthDomain → CoreDomain 으로 승격.
  CoreDomain/UMCFoundation 타입에만 의존해 승격 비용이 없고, Home → AuthDomain 은
  의존 방향이 잘못됐다. 기존 Auth 소비처는 이미 CoreDomain 을 import 하고 있어 수정 0건
- 홈 프로필 로드 성공 시 저장 동기화 → 최근 공지 조회 순으로 고정.
  세션 중 역할·기수 변경과 FCM 토큰 재동기화가 홈에서도 반영된다
- roles 에 기수가 없을 때의 공지 gisuId 폴백 복원

일정 아이콘 분류
- 분류 완료 전 .general 로 폴백해 아이콘이 팝인하던 문제 수정 — 선택 날짜 범위만 게이팅
- 월 전체 순차 for await 를 TaskGroup 병렬로 전환
- stale 가드를 분류뿐 아니라 fetchSchedules 응답 반영까지 한 세대 번호로 확장.
  늦게 도착한 이전 달 응답이 최신 분류와 어긋나 게이트가 영구 로딩으로 굳는 것을 막는다

- SeasonCard 에 Equatable 복원 (레이아웃 무변경)

* 🐛[Fix] 상벌점 소수점이 0으로 절삭되던 문제 수정

PointLog.point 가 Int 인데 원본 ProfileChallengerPoint.point 는 Double(decodeDoubleFlexible)
이라, 매핑에서 Int(point.point) 로 절삭돼 -0.5 점이 조용히 0 이 됐다.

- PointLog.point 와 HomeGeneration 의 penaltyPoint/rewardPoint 를 Double 로 통일하고
  매핑의 Int 변환 제거. 합계도 Double 로 누적
- 절대 규칙 #2 는 "서버 응답 정수"의 String 통일 규칙이라 소수 값인 이 필드는 대상이 아니다.
  판단 근거를 PointLog 프로퍼티 주석에 남김
- 표시는 .number.precision(.fractionLength(0...1)) 로 2.0 → "2", -0.5 → "-0.5"
- 회귀 테스트 fetchMyProfileKeepsFractionalPoints 추가

* 🔧[Chore] Preview #if DEBUG 가드 전역 스윕

절대 규칙 #5 에 따라 Preview 본문과 그 안의 목/픽스처 데이터가 릴리스 빌드에 포함되지
않도록 미가드 Preview 전체를 #if DEBUG 로 감쌌다.

- Home 11 · Notice 11 · CoreUIComponents 5 · Activity 2
- 저장소 전역 미가드 Preview 잔여 0

* 📝[Docs] Home 마이그레이션 잔여 검토 문서 추가 및 매핑표 동기화

- home-migration-plan.md 재작성. 초판의 그룹 A~I 는 전부 머지돼 stale 이었고,
  실제 잔여는 미이식 파일이 아니라 결선되지 않은 파이프라인·화면 내 누락·규약 부채였다.
  P0 5건 / P1 7건 / P2 5건과 처리 현황, 원안에서 달라진 판단, 후속 잔여를 정리
- tuist-file-mapping.md Home 구간 32행 동기화.
  Activity/Core/Notice 로 간 21행의 목적지 정정, 레이어 오기 2건, 개명 주석 3건,
  dead/superseded 마커 보강. 참조가 0이 아닌 항목은 범례 정의에 맞춰 dead 가 아닌
  superseded 로 표기

+++ UMCApp/Features/Home/Domain/Sources/Models/HomeGeneration.swift
@@ -17,8 +17,9 @@ public struct HomeGeneration: Identifiable, Equatable, Sendable {
 
     public let gisuId: String
     public let gen: String
-    public let penaltyPoint: Int
-    public let rewardPoint: Int
+    /// ``PointLog/point`` 합계이므로 소수 배점을 잃지 않도록 `Double`로 유지한다.
+    public let penaltyPoint: Double
+    public let rewardPoint: Double
     public let pointLogs: [PointLog]
 
     // MARK: - Init
@@ -26,8 +27,8 @@ public struct HomeGeneration: Identifiable, Equatable, Sendable {
     public init(
         gisuId: String,
         gen: String,
-        penaltyPoint: Int,
-        rewardPoint: Int,
+        penaltyPoint: Double,
+        rewardPoint: Double,
         pointLogs: [PointLog]
     ) {
         self.gisuId = gisuId
